requirements.txt
annotated-types==0.7.0
bcrypt==5.0.0
boto3==1.42.32
botocore==1.42.32
ecdsa==0.19.1
jmespath==1.0.1
pyasn1==0.6.2
pydantic[email]==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
rsa==4.9.1
s3transfer==0.16.0
six==1.17.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
coverage==7.4.1
pytest==9.0.2
passlib==1.7.4